# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Umrah Travel ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /flight-quotations:
    get:
      operationId: listFlightQuotations
      tags: [flights]
      summary: List flight quotations
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlightQuotation"
    post:
      operationId: createFlightQuotation
      tags: [flights]
      summary: Create flight quotation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlightQuotationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlightQuotation"
  /flight-quotations/{id}:
    get:
      operationId: getFlightQuotation
      tags: [flights]
      summary: Get flight quotation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlightQuotation"
    patch:
      operationId: updateFlightQuotation
      tags: [flights]
      summary: Update flight quotation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlightQuotationUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlightQuotation"
    delete:
      operationId: deleteFlightQuotation
      tags: [flights]
      summary: Delete flight quotation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── ACCOUNTING ────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── AUTH ──────────────────────────────────────────────────────
...
    FlightQuotation:
      type: object
      required: [id, clientId, clientName, origin, destination, departureDate, passengers, status, amount, currency, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        origin:
          type: string
        destination:
          type: string
        departureDate:
          type: string
        returnDate:
          type: ["string", "null"]
        passengers:
          type: integer
        cabinClass:
          type: string
          enum: [economy, business, first]
        airline:
          type: ["string", "null"]
        flightNumber:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, sent, booked, cancelled]
        amount:
          type: number
        currency:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    FlightQuotationInput:
      type: object
      required: [clientId, origin, destination, departureDate, passengers, cabinClass, amount, currency]
      properties:
        clientId:
          type: integer
        origin:
          type: string
        destination:
          type: string
        departureDate:
          type: string
        returnDate:
          type: string
        passengers:
          type: integer
        cabinClass:
          type: string
          enum: [economy, business, first]
        airline:
          type: string
        flightNumber:
          type: string
        amount:
          type: number
        currency:
          type: string
        notes:
          type: string
    FlightQuotationUpdate:
      type: object
      properties:
        status:
          type: string
        airline:
          type: string
        flightNumber:
          type: string
        amount:
          type: number
        notes:
          type: string

    # ── ACCOUNTING ─────────────────────────────────────────────────
...